spider/requirements.txt
py-postgresql==1.1.0
redis==2.7.1
